Unbreakable code
This is a bit same just the difference is that the secret text is encrypted multiple number of times as much as you like. Even a millions times !!. The cipher code and key is printed and it is literally not possible to break it unless the person lives for another 100 years or so.
Have fun nerds.